+++ shuup/admin/utils/bs3_renderers.py
@@ -10,6 +10,7 @@ from __future__ import unicode_literals
 from bootstrap3.renderers import FieldRenderer
 from bootstrap3.utils import add_css_class
 from django.forms import DateField, DateTimeField, ModelMultipleChoiceField
+from django.utils.html import escape, strip_tags
 
 
 class AdminFieldRenderer(FieldRenderer):
@@ -63,8 +64,14 @@ class AdminFieldRenderer(FieldRenderer):
         if self.field_errors:
             errors = "<br>".join(self.field_errors)
             html += '<div class="help-block error-block">{error}</div>'.format(error=errors)
-
         if self.field_help:
-            html += '<span class="help-block">{help}</span>'.format(help=self.field_help)
-
+            html += "<span class='help-popover-btn'>"
+            # tabindex is required for popover to function but we don't actually want to be able to tab to it
+            # so set a large tabindex
+            html += "<a class='btn' data-toggle='popover' data-placement='bottom' role='button' tabindex='50000' "
+            html += "data-trigger='focus' title='{title}' data-content='{help}'>".format(
+                title=self.field.label, help=escape(strip_tags(self.field_help)))
+            html += "<i class='fa fa-question-circle'></i>"
+            html += "</a>"
+            html += "</span>"
         return html