+++ doc/changelog.rst
@@ -20,6 +20,9 @@ Localization
 Admin
 ~~~~~
 
+- Show help text as popovers
+- Add admin walkthrough
+
 Addons
 ~~~~~~